Remove need to initialize DynamicView
This improves DynamicView interoperability by having DynamicView.dtype match np.ndarray.dtype (giving it all of the expected attributes of a numpy dtype). As a result of this refactoring, the dtype is stored in the underlying data which is now initialized as an empty array of the correct type, rather than None. This allows the DynamicView to immediately be in a valid state, without requiring separate initialization