Refactor to store slots separately from other states, and support passing in slot values
Since the slots are a hard dependency for *growing* other states, it's a bit clearer to just store them as a special attribute (like UIDs) rather than lumping them in with the other model states.

Also, it's important that as part of growing the People, the actual slot values are inserted *before* any other states are grown, since in principle any state could depend on RNGs and therefore on slots. So the grow() function now takes an optional argument for new_slots, to support demographic modules that optionally calculate slot values for new agents (e.g., Pregnancy) noting that not all modules do this (e.g., the basic birth module does not, because it doesn't assign births to specific agents)