Improve robustness to input type
Currently the isinstance check can fail if a numpy integer is passed in. By testing if it's a scalar, all integer types are supported and if a float is provided, a descriptive error will be raised by distribution.sample.

Also introduces a check for whether the max slot is INT_NAN - this could probably be removed later, but there have been some cases during development where uninitialized slots were inadvertantly used