Distribution auto-jumping bugfix
Distributions keep track of the number of jumps in a variable named `ind`. With auto-jumping, `ind` is increased by 1 each time the distribution is called. At the beginning of each time step, we were calling jump() with `to=self.ti+1`, which sets all distribution `ind` values to ti+1.

The major problem here is that with auto-jumping, a user could call a distribution multiple times per time step. The jumps counter `ind` might start at say 1, and subsequently increase to 2, 3, and 4 during a step (if a user calls `rvs` several times). But then at the next time step, step() was setting `ind` to 2, causing repeated random numbers to be used!

The easy one-line fix is to remove the `to=` argument to `jump` inside the `step` loop.

NOTE: This science bugfix could change simulation results.