Allow RNGs to jump to the same state
With auto=True by default (and anticipated to always be left as True by users) reuse of random numbers is unlikely. This case can be encountered if a Product contains Dist and is also used by multiple Interventions, in which case the RNG may be jumped to the current time multiple times. Although this does not maintain CRN, the results are still valid because the random numbers won't be reused, so this change allows this functionality rather than raising an error.